+++ lib/types/handlebars/index.ts
@@ -1,0 +1,5 @@
+export interface TemplateSpecification {
+    code?: string;
+    map?: string;
+    [key: string | number]: any;
+}