+++ lib/types/plugin-options/index.ts
@@ -1,0 +1,4 @@
+import { IPluginOptions } from "./params";
+
+export interface HandlebarsPluginOptions extends CompileOptions, IPluginOptions  {
+}