+++ lib/types/source-map/index.ts
@@ -1,0 +1,2 @@
+export type SourceData = [string, string]
+export type SourceDataMap = Map<string, string>