+++ tests/index.test.ts
@@ -1,5 +1,5 @@
 import type { Plugin } from 'rollup';
-import handlebarsCompilerPlugin from '../lib';
+import handlebarsCompilerPlugin from '../lib/index';
 
 describe('Test command', () => {
     it('this is the first test', () => {