+++ types/source-map-parser.d.ts
@@ -1,6 +1,0 @@
-import { SourceDataMap } from './types';
-export declare class SourceMapParser {
-    sourceMap: SourceDataMap;
-    constructor(sourceMap: SourceDataMap);
-    getFiles(directory: string, extname: string): string[];
-}